+++ cli/tests/runner.rs
@@ -53,6 +53,7 @@ mod test_revset_output;
 mod test_root;
 mod test_shell_completion;
 mod test_show_command;
+mod test_signature_templates;
 mod test_sparse_command;
 mod test_split_command;
 mod test_squash_command;