+++ cli/tests/test_log_command.rs
@@ -44,9 +44,12 @@ fn test_log_with_no_template() {
     - builtin_log_comfortable
     - builtin_log_compact
     - builtin_log_detailed
+    - builtin_log_detailed_with_sig
     - builtin_log_oneline
     - builtin_op_log_comfortable
     - builtin_op_log_compact
+    - builtin_sig_detailed
+    - builtin_sig_status
     - commit_summary_separator
     - description_placeholder
     - email_placeholder