+++ lib/src/mock_signing.rs
@@ -1,15 +1,33 @@
+// Copyright 2023 The Jujutsu Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+//! Generic APIs to work with cryptographic signatures created and verified by
+//! various backends.
+
 use hex::ToHex;
 use jj_lib::content_hash::blake2b_hash;
 use jj_lib::signing::{SigStatus, SignError, SignResult, SigningBackend, Verification};
 
 #[derive(Debug)]
-pub struct TestSigningBackend;
+#[allow(missing_docs)]
+pub struct MockSigningBackend;
 
-const PREFIX: &str = "--- JJ-TEST-SIGNATURE ---\nKEY: ";
+const PREFIX: &str = "--- JJ-MOCK-SIGNATURE ---\nKEY: ";
 
-impl SigningBackend for TestSigningBackend {
+impl SigningBackend for MockSigningBackend {
     fn name(&self) -> &str {
-        "test"
+        "mock"
     }
 
     fn can_read(&self, signature: &[u8]) -> bool {
@@ -37,18 +55,11 @@ impl SigningBackend for TestSigningBackend {
         let key = (!key.is_empty()).then_some(std::str::from_utf8(key).unwrap().to_owned());
 
         let sig = self.sign(data, key.as_deref())?;
-        if sig == signature {
-            Ok(Verification {
-                status: SigStatus::Good,
-                key,
-                display: None,
-            })
+        let status = if sig == signature {
+            SigStatus::Good
         } else {
-            Ok(Verification {
-                status: SigStatus::Bad,
-                key,
-                display: None,
-            })
-        }
+            SigStatus::Bad
+        };
+        Ok(Verification::new(status, key, None))
     }
 }