+++ lib/tests/test_signing.rs
@@ -1,9 +1,10 @@
+use assert_matches::assert_matches;
 use jj_lib::backend::{MillisSinceEpoch, Signature, Timestamp};
+use jj_lib::mock_signing::MockSigningBackend;
 use jj_lib::repo::Repo;
 use jj_lib::settings::UserSettings;
-use jj_lib::signing::{SigStatus, SignBehavior, Signer, Verification};
+use jj_lib::signing::{SignBehavior, Signer};
 use test_case::test_case;
-use testutils::test_signing_backend::TestSigningBackend;
 use testutils::{create_random_commit, write_random_commit, TestRepoBackend, TestWorkspace};
 
 fn user_settings(sign_all: bool) -> UserSettings {
@@ -33,20 +34,14 @@ fn someone_else() -> Signature {
     }
 }
 
-fn good_verification() -> Option<Verification> {
-    Some(Verification {
-        status: SigStatus::Good,
-        key: Some("impeccable".to_owned()),
-        display: None,
-    })
-}
+const GOOD_VERIFICATION: &str = r#"Ok(Some(Verification { status: Good, key: Some("impeccable"), display: None, backend: Some("mock") }))"#;
 
 #[test_case(TestRepoBackend::Local ; "local backend")]
 #[test_case(TestRepoBackend::Git ; "git backend")]
 fn manual(backend: TestRepoBackend) {
     let settings = user_settings(true);
 
-    let signer = Signer::new(Some(Box::new(TestSigningBackend)), vec![]);
+    let signer = Signer::new(Some(Box::new(MockSigningBackend)), vec![]);
     let test_workspace = TestWorkspace::init_with_backend_and_signer(&settings, backend, signer);
 
     let repo = &test_workspace.repo;
@@ -66,17 +61,17 @@ fn manual(backend: TestRepoBackend) {
     tx.commit("test");
 
     let commit1 = repo.store().get_commit(commit1.id()).unwrap();
-    assert_eq!(commit1.verification().unwrap(), good_verification());
+    assert_eq!(format!("{:?}", commit1.verification()), GOOD_VERIFICATION);
 
     let commit2 = repo.store().get_commit(commit2.id()).unwrap();
-    assert_eq!(commit2.verification().unwrap(), None);
+    assert_matches!(commit2.verification(), Ok(None));
 }
 
 #[test_case(TestRepoBackend::Git ; "git backend")]
 fn keep_on_rewrite(backend: TestRepoBackend) {
     let settings = user_settings(true);
 
-    let signer = Signer::new(Some(Box::new(TestSigningBackend)), vec![]);
+    let signer = Signer::new(Some(Box::new(MockSigningBackend)), vec![]);
     let test_workspace = TestWorkspace::init_with_backend_and_signer(&settings, backend, signer);
 
     let repo = &test_workspace.repo;
@@ -95,14 +90,14 @@ fn keep_on_rewrite(backend: TestRepoBackend) {
     let rewritten = mut_repo.rewrite_commit(&settings, &commit).write().unwrap();
 
     let commit = repo.store().get_commit(rewritten.id()).unwrap();
-    assert_eq!(commit.verification().unwrap(), good_verification());
+    assert_eq!(format!("{:?}", commit.verification()), GOOD_VERIFICATION);
 }
 
 #[test_case(TestRepoBackend::Git ; "git backend")]
 fn manual_drop_on_rewrite(backend: TestRepoBackend) {
     let settings = user_settings(true);
 
-    let signer = Signer::new(Some(Box::new(TestSigningBackend)), vec![]);
+    let signer = Signer::new(Some(Box::new(MockSigningBackend)), vec![]);
     let test_workspace = TestWorkspace::init_with_backend_and_signer(&settings, backend, signer);
 
     let repo = &test_workspace.repo;
@@ -132,7 +127,7 @@ fn manual_drop_on_rewrite(backend: TestRepoBackend) {
 fn forced(backend: TestRepoBackend) {
     let settings = user_settings(true);
 
-    let signer = Signer::new(Some(Box::new(TestSigningBackend)), vec![]);
+    let signer = Signer::new(Some(Box::new(MockSigningBackend)), vec![]);
     let test_workspace = TestWorkspace::init_with_backend_and_signer(&settings, backend, signer);
 
     let repo = &test_workspace.repo;
@@ -148,14 +143,14 @@ fn forced(backend: TestRepoBackend) {
     tx.commit("test");
 
     let commit = repo.store().get_commit(commit.id()).unwrap();
-    assert_eq!(commit.verification().unwrap(), good_verification());
+    assert_eq!(format!("{:?}", commit.verification()), GOOD_VERIFICATION);
 }
 
 #[test_case(TestRepoBackend::Git ; "git backend")]
 fn configured(backend: TestRepoBackend) {
     let settings = user_settings(true);
 
-    let signer = Signer::new(Some(Box::new(TestSigningBackend)), vec![]);
+    let signer = Signer::new(Some(Box::new(MockSigningBackend)), vec![]);
     let test_workspace = TestWorkspace::init_with_backend_and_signer(&settings, backend, signer);
 
     let repo = &test_workspace.repo;
@@ -167,5 +162,5 @@ fn configured(backend: TestRepoBackend) {
     tx.commit("test");
 
     let commit = repo.store().get_commit(commit.id()).unwrap();
-    assert_eq!(commit.verification().unwrap(), good_verification());
+    assert_eq!(format!("{:?}", commit.verification()), GOOD_VERIFICATION);
 }